update delay arg to accept floats, handle crashes, refactor

+++ args.go
@@ -8,7 +8,7 @@ import (
 
 var (
 	initialVolume  = flag.Float64("volume", 10, "Initial volume level (0 to 100)")
-	initialDelay   = flag.Int64("delay", 5, "Initial delay in milliseconds between operations")
+	initialDelay   = flag.Float64("delay", 5, "Initial delay in milliseconds between operations")
 	initialSize    = flag.Int("size", 100, "Initial array size")
 	initialSort    = flag.String("sort", "quick", "Initial sorting algorithm")
 	fpsFlag        = flag.Int("fps", 60, "Frames per second for rendering")
@@ -23,8 +23,8 @@ func parseArgs() bool {
 	if *initialVolume < 0.0 || *initialVolume > 1.0 {
 		*initialVolume = 0.1
 	}
-	if *initialDelay < 0 {
-		*initialDelay = 5
+	if *initialDelay < 0.0 {
+		*initialDelay = 5.0
 	}
 	if *fpsFlag <= 0 {
 		*fpsFlag = 60
